SSLUtils: move tests and update due to JDK-8211883
